+++ CHANGELOG.md
@@ -14,6 +14,8 @@ All notable changes to this project will be documented in this file.
 - API: Add `cosmicLink.verifySep7()`. Verify SEP-0007 signature by resolving
   `cosmicLink.extra.domain`, if any. Throw an error if the signature is not
   valid.
+- API: Add `cosmicLink.signSep7()`. Parameters: domain, keypair. Sign SEP-0007
+  link for **domain**, using **keypair**.
 
 ## 1.8.1 - 2019-09-02